A lot of fixes and graphical changes -- 98% "finished state"

+++ joro.too.Web/Models/ViewMovieModel.cs
@@ -9,7 +9,7 @@ public class ViewMovieModel
     public string description { get; set; }
     public string imgsrc { get; set; }
     public List<Genre> genres { get; set; }
-    public decimal rating { get; set; }
+    public float rating { get; set; }
     public List<ActorInGivenMediaModel> actors { get; set; }
     public VideoViewModel? movie;
